+++ hello-graphql-rails/db/migrate/20180410083626_create_likes.rb
@@ -1,8 +1,8 @@
 class CreateLikes < ActiveRecord::Migration[5.2]
   def change
     create_table :likes do |t|
-      t.integer :post_id
-      t.integer :user_id
+      t.bigint :post_id
+      t.bigint :user_id
 
       t.timestamps
     end
